requirements.txt
Flask==1.0.2
requests==2.18.4
gunicorn
python-dotenv

